Add saving in Playerprefs, persistance
Coins and upgrades are now stored from one session to the next. basic crit dmg implemented, need to add/fix buying it

+++ Assets/Scripts/Health.cs
@@ -41,7 +41,7 @@ public class Health : MonoBehaviour
                 UnityEngine.Debug.Log("Crit");
             }*/
 
-            damage = damage * (100 / (100 + armor));
+            damage = (int)((float)damage * (100.0 / (100.0 + (float)armor)));
             if (damage <= 1)
             {
                 damage = 1;
